Migrate layout module to have a schema
This is the mechanism by which it will be consumed downstream, but it's also useful for testing the forthcoming implementation of schema dependencies.
