+++ redwood-layout-api/gradle.properties
@@ -1,0 +1,2 @@
+POM_ARTIFACT_ID=redwood-layout-api
+POM_NAME=Redwood Layout API
