+++ redwood-layout-schema/compose/gradle.properties
@@ -1,0 +1,2 @@
+POM_ARTIFACT_ID=redwood-layout-compose
+POM_NAME=Redwood Layout Composables
