+++ redwood-layout-schema/gradle.properties
@@ -1,0 +1,2 @@
+POM_ARTIFACT_ID=redwood-layout-schema
+POM_NAME=Redwood Layout Schema
