+++ redwood-layout-schema/layout-modifiers/build.gradle
@@ -1,0 +1,21 @@
+apply plugin: 'org.jetbrains.kotlin.multiplatform'
+apply plugin: 'app.cash.redwood.generator.layoutmodifiers'
+apply plugin: 'com.vanniktech.maven.publish'
+apply plugin: 'org.jetbrains.dokka' // Must be applied here for publish plugin.
+
+kotlin {
+  apply from: "${rootDir}/addAllTargets.gradle"
+
+  sourceSets {
+    commonMain {
+      dependencies {
+        api projects.redwoodLayoutApi
+      }
+    }
+  }
+}
+
+redwoodSchema {
+  source = projects.redwoodLayoutSchema
+  type = 'app.cash.redwood.layout.RedwoodLayout'
+}
