+++ redwood-layout-schema/layout-modifiers/gradle.properties
@@ -1,0 +1,2 @@
+POM_ARTIFACT_ID=redwood-layout-modifiers
+POM_NAME=Redwood Layout Modifiers
