+++ redwood-layout-schema/src/main/kotlin/app/cash/redwood/layout/modifiers.kt
@@ -1,0 +1,43 @@
+/*
+ * Copyright (C) 2022 Square, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package app.cash.redwood.layout
+
+import app.cash.redwood.layout.api.CrossAxisAlignment
+import app.cash.redwood.schema.LayoutModifier
+
+/** Grow the node relative to [value] along the main axis. */
+@LayoutModifier(1_000_001)
+public data class GrowLayoutModifier(
+  val value: Int,
+)
+
+/** Shrink the node relative to [value] along the main axis. */
+@LayoutModifier(1_000_002)
+public data class ShrinkLayoutModifier(
+  val value: Int,
+)
+
+/** Set the alignment for a node along the horizontal axis. */
+@LayoutModifier(1_000_003)
+public data class HorizontalAlignmentLayoutModifier(
+  val alignment: CrossAxisAlignment,
+)
+
+/** Set the alignment for a node along the vertical axis. */
+@LayoutModifier(1_000_004)
+public data class VerticalAlignmentLayoutModifier(
+  val alignment: CrossAxisAlignment,
+)
