+++ redwood-layout-schema/widget/gradle.properties
@@ -1,0 +1,2 @@
+POM_ARTIFACT_ID=redwood-layout-widget
+POM_NAME=Redwood Layout Widgets
