+++ redwood-layout-view/README.md
@@ -1,0 +1,3 @@
+# Redwood Layout
+
+This artifact includes `Row` and `Column` widget implementations for Android Views.
