+++ redwood-layout-view/build.gradle
@@ -1,0 +1,16 @@
+apply plugin: 'com.android.library'
+apply plugin: 'org.jetbrains.kotlin.android'
+apply plugin: 'org.jetbrains.kotlin.plugin.serialization'
+apply plugin: 'app.cash.paparazzi'
+apply plugin: 'com.vanniktech.maven.publish'
+apply plugin: 'org.jetbrains.dokka' // Must be applied here for publish plugin.
+
+dependencies {
+  implementation projects.redwoodFlexContainer
+  api projects.redwoodLayoutSchema.widget
+  implementation libs.androidx.core
+}
+
+android {
+  namespace 'app.cash.redwood.layout.view'
+}
