+++ redwood-layout-view/gradle.properties
@@ -1,0 +1,2 @@
+POM_ARTIFACT_ID=redwood-layout-view
+POM_NAME=Redwood Layout for Android Views
