+++ redwood-layout-view/src/main/kotlin/app/cash/redwood/layout/ViewLayoutWidgetFactory.kt
@@ -1,0 +1,29 @@
+/*
+ * Copyright (C) 2022 Square, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package app.cash.redwood.layout
+
+import android.content.Context
+import android.view.View
+import app.cash.redwood.layout.widget.Column
+import app.cash.redwood.layout.widget.RedwoodLayoutWidgetFactory
+import app.cash.redwood.layout.widget.Row
+
+public class ViewRedwoodLayoutWidgetFactory(
+  private val context: Context,
+) : RedwoodLayoutWidgetFactory<View> {
+  public override fun Column(): Column<View> = ViewColumn(context)
+  public override fun Row(): Row<View> = ViewRow(context)
+}
