+++ redwood-layout-view/src/test/snapshots/images/app.cash.redwood.layout_ViewLayoutTest_rowCenter.png
