--- redwood-layout/build.gradle
@@ -1,31 +1,0 @@
-apply plugin: 'org.jetbrains.kotlin.multiplatform'
-apply plugin: 'org.jetbrains.kotlin.plugin.serialization'
-apply plugin: 'com.android.library'
-apply plugin: 'app.cash.paparazzi'
-apply plugin: 'com.vanniktech.maven.publish'
-apply plugin: 'org.jetbrains.dokka' // Must be applied here for publish plugin.
-
-kotlin {
-  apply from: "${rootDir}/addAllTargets.gradle"
-
-  sourceSets {
-    commonMain {
-      dependencies {
-        implementation projects.redwoodFlexContainer
-        api projects.redwoodRuntime
-        api projects.redwoodWidget
-        api libs.jetbrains.compose.runtime
-        api libs.kotlinx.serialization.core
-      }
-    }
-    androidMain {
-      dependencies {
-        implementation libs.androidx.core
-      }
-    }
-  }
-}
-
-android {
-  namespace 'app.cash.redwood.layout'
-}
