--- redwood-layout/gradle.properties
@@ -1,2 +1,0 @@
-POM_ARTIFACT_ID=redwood-layout
-POM_NAME=Redwood Layout
