--- redwood-layout/src/commonMain/kotlin/app/cash/redwood/layout/utils.kt
@@ -1,43 +1,0 @@
-/*
- * Copyright (C) 2022 Square, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-@file:JvmName("common_utils")
-
-package app.cash.redwood.layout
-
-import app.cash.redwood.flexcontainer.AlignItems
-import app.cash.redwood.flexcontainer.JustifyContent
-import app.cash.redwood.flexcontainer.Spacing
-import kotlin.jvm.JvmName
-
-internal fun MainAxisAlignment.toJustifyContent() = when (this) {
-  MainAxisAlignment.Start -> JustifyContent.FlexStart
-  MainAxisAlignment.Center -> JustifyContent.Center
-  MainAxisAlignment.End -> JustifyContent.FlexEnd
-  MainAxisAlignment.SpaceBetween -> JustifyContent.SpaceBetween
-  MainAxisAlignment.SpaceAround -> JustifyContent.SpaceAround
-  MainAxisAlignment.SpaceEvenly -> JustifyContent.SpaceEvenly
-  else -> throw AssertionError()
-}
-
-internal fun CrossAxisAlignment.toAlignItems() = when (this) {
-  CrossAxisAlignment.Start -> AlignItems.FlexStart
-  CrossAxisAlignment.Center -> AlignItems.Center
-  CrossAxisAlignment.End -> AlignItems.FlexEnd
-  CrossAxisAlignment.Stretch -> AlignItems.Stretch
-  else -> throw AssertionError()
-}
-
-internal fun Padding.toSpacing() = Spacing(start, end, top, bottom)
