+++ settings.gradle
@@ -4,7 +4,12 @@ include ':redwood-cli'
 include ':redwood-compose'
 include ':redwood-flex-container'
 include ':redwood-gradle-plugin'
-include ':redwood-layout'
+include ':redwood-layout-api'
+include ':redwood-layout-schema'
+include ':redwood-layout-schema:compose'
+include ':redwood-layout-schema:layout-modifiers'
+include ':redwood-layout-schema:widget'
+include ':redwood-layout-view'
 include ':redwood-protocol'
 include ':redwood-protocol-compose'
 include ':redwood-protocol-widget'
